feat: redesign layout with additional sections
Redesigned the layout from the original The Odin Project Landing Page.
Modified the hero section and replaced the image.
Replaced the random information section with a proper services section.
Removed the qoutes section, and  restructured the CTA section as contacts section.
Added images for visual enhancement. Updated HTML and CSS files.
Portfolio and contacts section are still in progress.